requirements.txt
boto3
pandas>=0.22
sqlalchemy>=1.0
biom-format
psycopg2
XlsxWriter
gitpython
